Improve step definition locator
With this patch find_step.rb script will use gherkin, the library to
parse the cucumber features. It will allow us to extract additional
information about step on the cursor. Like if it has examples defined.
For example, trying to locate definition of the step with placeholder
like `And "John" user <two> be shown` will fail with previous
implementation

  Scenario Outline: Search by gender
    Given I open search page
    And I fill in gender as "<text>"
    Then "Jane" user <one> be shown
    And "John" user <two> be shown
    Examples:
      | text  | one        | two        |
      | man   | should not | should     |
      | woman | should     | should not |

With this patch the user will be prompted (using IDO) with two choices:
`"John" user should be shown` and `"John" user should not be shown`,
basically it will be able to jump to any valid step which cucumber
understand