Use line number instead of scenario name when run it
Line number is more robust because cucumber does pattern matching for
'-n' argument, so if you happened to have scenario, which name is
happened to be part of another scenario in the same file, feature-mode
will run them both. Another case is that it is not possible to run on
specific example of scenario outline. With line number argument, it is
possible, just set cursor to the row in the examples table.