Search project installed Gems for step definitions
Some projects will define extra cucumber step definitions in separate
installable Gems. In this scenario the goto-step-definition capability
will need to also search those gems for the step definitions.

This feature introduces feature-step-search-gems-path to indicate an
additional location to search for steps. Implication is that someone
would use bundler to install required gems and use the '--path ./gems'
to do so typically.